renames files containing workers
